lazy load post because will be moving this out soon
